fix: correct functions

+++ ft_strlen.c
@@ -1,8 +1,8 @@
 #include <ft_lib.h>
 
-int ft_strlen(const char *s) {
+size_t ft_strlen(const char *s) {
 
-  int count = 0;
+  size_t count = 0;
   while (*s != '\0') {
     count++;
     s++;
